requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
jupyter>=1.0.0
pyyaml>=6.0

# Deep Learning (optional - add when needed)
# torch>=2.0.0
# torchvision>=0.15.0
# transformers>=4.30.0
# sentence-transformers>=2.2.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.3.0

# Development
ipykernel>=6.25.0
ipywidgets>=8.0.0
